chore(ui/message): convert message image modal to radix ui
