+++ packages/ui/package.json
@@ -25,6 +25,7 @@
 		"@answeroverflow/hooks": "*",
 		"@heroicons/react": "^2.0.13",
 		"@next/font": "^13.1.2",
+		"@radix-ui/react-alert-dialog": "^1.0.3",
 		"@radix-ui/react-avatar": "^1.0.2",
 		"@radix-ui/react-dropdown-menu": "^2.0.4",
 		"@radix-ui/react-label": "^2.0.1",
