+++ packages/ui/src/components/pages/MessageResultPage.tsx
@@ -12,8 +12,6 @@ import {
 } from '../primitives';
 import { MessagesSearchBar } from './SearchPage';
 import { useTrackEvent } from '@answeroverflow/hooks';
-import { MessageModal } from '../primitives/MessageModal';
-import { useState } from 'react';
 export type MessageResultPageProps = {
 	messages: APIMessageWithDiscordAccount[];
 	server: ServerPublic;
@@ -31,10 +29,6 @@ export function MessageResultPage({
 	thread,
 }: MessageResultPageProps) {
 	const isUserInServer = useIsUserInServer(server.id);
-	const [isMessageModalOpen, setIsMessageModalOpen] = useState(false);
-	const [attatchmentSelected, setAttachmentSelected] = useState<
-		APIMessageWithDiscordAccount['attachments'][number] | undefined | null
-	>(null);
 
 	const firstMessage = messages.at(0);
 	if (!firstMessage) throw new Error('No message found'); // TODO: Handle this better
@@ -88,8 +82,6 @@ export function MessageResultPage({
 				message={message}
 				fullRounded
 				Blurrer={(props) => <MultiMessageBlurrer {...props} count={count} />}
-				setAttachment={setAttachmentSelected}
-				setOpen={setIsMessageModalOpen}
 			/>
 		);
 
@@ -117,11 +109,7 @@ export function MessageResultPage({
 				title={`${channelName} - ${server.name}`}
 				server={server}
 			/>
-			<MessageModal
-				isOpen={isMessageModalOpen}
-				setIsOpen={setIsMessageModalOpen}
-				attachment={attatchmentSelected}
-			/>
+
 			<div className="my-8 flex flex-col items-center justify-between gap-2 sm:flex-row sm:py-0">
 				<MessagesSearchBar />
 				<div className="shrink-0 sm:pl-8">
